+++ R/glycan-grob.R
@@ -18,7 +18,7 @@ glycanGrob <- function(
   structure,
   ...,
   show_linkage = TRUE,
-  orient = c("H", "V"),
+  orient = c("left", "right", "up", "down"),
   fuc_orient = c("flex", "up"),
   red_end = "",
   edge_linewidth = 0.8,